+++ packages/stream_video_flutter/lib/src/call_controls/controls/toggle_camera_option.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: deprecated_member_use_from_same_package
+
 import 'package:flutter/material.dart';
 
 import '../../../stream_video_flutter.dart';
@@ -9,7 +11,7 @@ class ToggleCameraOption extends StatelessWidget {
   const ToggleCameraOption({
     super.key,
     required this.call,
-    required this.localParticipant,
+    this.localParticipant,
     this.enabledCameraIcon = Icons.videocam_rounded,
     this.disabledCameraIcon = Icons.videocam_off_rounded,
     this.enabledCameraIconColor,
@@ -22,7 +24,8 @@ class ToggleCameraOption extends StatelessWidget {
   final Call call;
 
   /// The current local participant.
-  final CallParticipantState localParticipant;
+  /// If provided this [localParticipant] will be used, otherwise the localParticipant of the [call] will be used.
+  final CallParticipantState? localParticipant;
 
   /// The icon that is shown when the camera is enabled.
   final IconData enabledCameraIcon;
@@ -44,17 +47,26 @@ class ToggleCameraOption extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final enabled = localParticipant.isVideoEnabled;
+    Widget buildContent(bool enabled) {
+      return CallControlOption(
+        icon: enabled ? Icon(enabledCameraIcon) : Icon(disabledCameraIcon),
+        iconColor: enabled ? enabledCameraIconColor : disabledCameraIconColor,
+        backgroundColor: enabled
+            ? enabledCameraBackgroundColor
+            : disabledCameraBackgroundColor,
+        onPressed: () {
+          call.setCameraEnabled(enabled: !enabled);
+        },
+      );
+    }
 
-    return CallControlOption(
-      icon: enabled ? Icon(enabledCameraIcon) : Icon(disabledCameraIcon),
-      iconColor: enabled ? enabledCameraIconColor : disabledCameraIconColor,
-      backgroundColor: enabled
-          ? enabledCameraBackgroundColor
-          : disabledCameraBackgroundColor,
-      onPressed: () {
-        call.setCameraEnabled(enabled: !enabled);
-      },
+    if (localParticipant != null) {
+      return buildContent(localParticipant!.isVideoEnabled);
+    }
+    return PartialCallStateBuilder(
+      call: call,
+      selector: (state) => state.localParticipant?.isVideoEnabled ?? false,
+      builder: (_, enabled) => buildContent(enabled),
     );
   }
 }